+++ Source/Minecraft/Private/World/Generator/HeightGenerator.cpp
@@ -6,7 +6,7 @@
 void HeightGenerator::GenerateHeight(AChunk* Chunk)
 {
 	const FVector ChunkLocation = Chunk->GetActorLocation();
-	TArray<uint8>& HeightMap = Chunk->GetHeightMap();
+	TArray<int32>& HeightMap = Chunk->GetHeightMap();
 
 	for (int32 X = 0; X < CHUNK_SIZE; ++X)
 	{
